Consolidated date range and vol from VaR functions. Cleaned up doc strings further. Changed OLS output to OLSResults.
